Fixed signup and added some login functionality
