+++ views/sign-up.hbs
@@ -1,8 +1,8 @@
 <h2>sign-up</h2>
 
-<form action="sign-up" method="POST">
-    <input type="text" name="username" id="Username" placeholder="type in username"> 
-    <input type="password" name="password"id="password" placeholder="type in password">
+<form action="/sign-up" method="POST">
+    <input type="text" name="username" id="Username" placeholder="type in username">
+    <input type="password" name="password" id="password" placeholder="type in password">
     <input type="email" name="email" id="email" placeholder="type in your e-mail">
     <input type="email" name="emailRepetition" id="emailRepetition" placeholder="repeate your email">
     <button type="submit">create account</button>
